Added fully-integrated scenario to test on
To help guide and focus development, I've built the first integrated test. It is
a relatively simple test (building a loan repayment model) but is complicated
enough that it encapsulates some of the interesting behaviour of the package.

There is still a need for (1) a simpler case - such as a budget - and (2) a
more complicated case, such as a simple FMCG financial model.