Added class, method, directory structure
This commit - although it seems small - is actually really important. It has
built out the structure of the classes, methods and directory such that the
current test scenario "works" and doesn't crash upon executing. It doesn't do
anything (yet) but it means that code can be run and interpreted correctly.