Added some basic documentation
Although just auto-populated, and basic, information this commit serves as the
basis of PyFinancial's documentation. It has been set up that the docs are
automatically updated based on the docstrings of the modules so it helps to
write well-documented code. In future commits there will be accompanying text
with the package to make it even easier to use.